build: temporarily remove all bytecode
Fix reproducibility issue when building the AppImages:

Bytecode generated by cp38, cp39 and cp310 seems to be undeterministic,
even if the PYTHONHASHSEED=0 environment variable is set. This is
especially a problem when the missing bytecode of the stdlib in the
docker images gets compiled when running pip and installing deps.

All attempts at fixing this by running the compileall module afterwards
failed, as well as suppressing writing bytecode via the -B flag or the
PYTHONDONTWRITEBYTECODE=1 env var, so the only solution right now
is removing any bytecode.

When executing the AppImages and mounting the squashfs, bytecode can't
be written due to the nature of AppImages, which are read-only, so this
all comes at the cost of a slight performance degradation.